Preserve IDs & class names of unwrapped DIVs
Some documents use DIVs instead of P elements, or they wrap a single P element.
In both of these cases, readability will replace the DIV with P, but in the
process the ID of the original DIV would be discarded. This would break the
table of contents if it relied on that ID to link to that section.

This change makes it so that the "id" and "class" attributes are preserved when
a DIV has been swapped out with a P.