Test_parser: make failures for text content mismatches more readable
When Test_parser iterates over the myriad test pages in the suite, if any parsed
text for an article doesn't match what's in the `expected.html` fixture, the test
will generate a failure message using DiffPrettyText from diffmatchpatch.

This diff formatter will output the complete text with the additions and removals
marked up with ANSI escape sequences for terminal colors green and red. While this
is alright for shorter text, some of the test pages are really long, and even
when the mismatch is with just one word or one line, the entire text will always
be printed together with the test failure, making the overall Test_parser result
much harder to scroll through and comprehend.

This change replaces the test failure message with a diff generated in a way that
long passages of text are truncated with `<...>`. Furthermore, this adds diff
markers `{+ +}` for additions and `[- -]` for removals, the same markers as with
word diff mode from git-diff(1).