Optimize traversing the DOM when analyzing text content
Previously, this was an often-repeated construct in readability implementation:

    charCount(ps.getInnerText(node, true)) == 0

What this would do is:

- Call `dom.TextContent(node)` to append the contents of all individual text nodes together;
- Pass the result through `strings.TrimSpace`;
- Pass the result through the NormalizeSpaces regex which squashes consecutive runs of whitespace;
- Count the Unicode runes of the result;
- Finally, if the count is zero, the element would be considered "empty".

The above is an example of an incredibly costly operation that could be done much more efficiently, for example: walk the DOM subtree until the first non-space character is found, then bail out and conclude that the element has content. This barely needs any memory allocations, and is the approach taken in this PR to address a variety of counting or detecting tasks that share a similar purpose.

Benchmark before vs. after for processing a large HTML document reveals significant saving in memory allocations:

variant | times | ns/op      | Bytes/op   | allocs/op
--------|-------|------------|------------|----------
before  | 30    | 38,986,203 | 59,623,683 | 199,876
after   | 36    | 31,910,769 | 11,449,004 | 119,810